不要なコンポーネント削除
