--- src/components/organisms/SignInForm/index.ts
@@ -1,1 +1,0 @@
-export { SigninForm } from "./SigninForm"
