+++ src/components/pages/Signin/Signin.tsx
@@ -1,9 +1,20 @@
 import React, { FC } from "react"
-import { SigninForm } from "../../organisms/SignInForm"
+import firebase from "../../../services/Firebase";
+import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth"
 
+const uiConfig = {
+  signInFlow: 'popup',
+  signInSuccessUrl: "/",
+  signInOptions: [
+    firebase.auth.GoogleAuthProvider.PROVIDER_ID,
+    // firebase.auth.FacebookAuthProvider.PROVIDER_ID,
+    firebase.auth.EmailAuthProvider.PROVIDER_ID,
+    // firebaseui.auth.AnonymousAuthProvider.PROVIDER_ID
+  ],
+}
 
 export const Signin: FC = () => {
   return (
-    <SigninForm />
+    <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
   )
 }
